removing debugging lines
These lines were debugging a scenario where we were no longer authed
but did not receive a 401 as part of that auth.

This issue was resolved once we started using the Authorization header
instead of the ATC-Authorization cookie.

Concourse behaves much better to this gem now that it considers us an
api client instead of a CSRF lockdown